refactor: channel can use custom serializer
