+++ packages/connection/__test__/common/fury-extends/one-of.test.ts
@@ -1,13 +1,11 @@
 /* eslint-disable no-console */
 
-import {
-  OpenMessage,
-  PingMessage,
-  PongMessage,
-  ServerReadyMessage,
-  parse,
-  stringify,
-} from '../../../src/common/ws-channel';
+import { furySerializer } from '@opensumi/ide-connection';
+
+import { OpenMessage, PingMessage, PongMessage, ServerReadyMessage } from '../../../lib';
+
+const parse = furySerializer.deserialize;
+const stringify = furySerializer.serialize;
 
 describe('oneOf', () => {
   function testIt(obj: any) {
@@ -23,7 +21,6 @@ describe('oneOf', () => {
   it('should serialize and deserialize', () => {
     const obj = {
       kind: 'ping',
-      clientId: '123',
       id: '456',
     } as PingMessage;
 
@@ -31,7 +28,6 @@ describe('oneOf', () => {
 
     const obj2 = {
       kind: 'pong',
-      clientId: '123',
       id: '456',
     } as PongMessage;
