+++ packages/connection/src/common/serializer/types.ts
@@ -1,0 +1,4 @@
+export interface ISerializer<FROM, TO> {
+  serialize(data: FROM): TO;
+  deserialize(data: TO): FROM;
+}
