+++ packages/connection/src/common/ws-channel.ts
@@ -1,5 +1,3 @@
-import { Type } from '@furyjs/fury';
-
 import { EventEmitter } from '@opensumi/events';
 import {
   DisposableCollection,
@@ -9,51 +7,10 @@ import {
   randomString,
 } from '@opensumi/ide-core-common';
 
+import { ChannelMessage, ErrorMessageCode } from './channel/types';
 import { IConnectionShape } from './connection/types';
-import { oneOf } from './fury-extends/one-of';
 import { ISumiConnectionOptions, SumiConnection } from './rpc/connection';
 import { ILogger } from './types';
-
-/**
- * `ping` and `pong` are used to detect whether the connection is alive.
- */
-export interface PingMessage {
-  kind: 'ping';
-  id: string;
-  clientId: string;
-}
-
-/**
- * when server receive a `ping` message, it should reply a `pong` message, vice versa.
- */
-export interface PongMessage {
-  kind: 'pong';
-  id: string;
-  clientId: string;
-}
-
-/**
- * `data` message indicate that the channel has received some data.
- * the `content` field is the data, it should be a string.
- */
-export interface DataMessage {
-  kind: 'data';
-  id: string;
-  content: string;
-}
-
-export interface BinaryMessage {
-  kind: 'binary';
-  id: string;
-  binary: Uint8Array;
-}
-
-export interface CloseMessage {
-  kind: 'close';
-  id: string;
-  code: number;
-  reason: string;
-}
 
 export interface IWSChannelCreateOptions {
   /**
@@ -80,25 +37,23 @@ export class WSChannel {
 
   protected onBinaryQueue = this._disposables.add(new EventQueue<Uint8Array>());
 
-  protected sendQueue: Uint8Array[] = [];
+  protected sendQueue: ChannelMessage[] = [];
   protected _isServerReady = false;
   protected _ensureServerReady: boolean | undefined;
 
   public id: string;
 
   public channelPath: string;
+  public clientId: string;
+
+  protected LOG_TAG = '[WSChannel]';
 
   logger: ILogger = console;
 
-  static forClient(connection: IConnectionShape<Uint8Array>, options: IWSChannelCreateOptions) {
+  static forClient(connection: IConnectionShape<ChannelMessage>, options: IWSChannelCreateOptions) {
     const disposable = new DisposableCollection();
     const channel = new WSChannel(connection, options);
-
-    disposable.push(
-      connection.onMessage((data) => {
-        channel.dispatch(parse(data));
-      }),
-    );
+    disposable.push(channel.listen());
 
     connection.onceClose(() => {
       disposable.dispose();
@@ -107,10 +62,10 @@ export class WSChannel {
     return channel;
   }
 
-  constructor(public connection: IConnectionShape<Uint8Array>, options: IWSChannelCreateOptions) {
+  constructor(public connection: IConnectionShape<ChannelMessage>, options: IWSChannelCreateOptions) {
     const { id, logger, ensureServerReady } = options;
     this.id = id;
-
+    this.LOG_TAG = `[WSChannel id=${this.id}]`;
     if (logger) {
       this.logger = logger;
     }
@@ -120,7 +75,7 @@ export class WSChannel {
     this._disposables.add(this.emitter.on('binary', (data) => this.onBinaryQueue.push(data)));
   }
 
-  protected inqueue(data: Uint8Array) {
+  protected inqueue(data: ChannelMessage) {
     if (this._ensureServerReady && !this._isServerReady) {
       if (!this.sendQueue) {
         this.sendQueue = [];
@@ -182,6 +137,20 @@ export class WSChannel {
       case 'binary':
         this.emitter.emit('binary', msg.binary);
         break;
+      case 'error':
+        this.logger.error(this.LOG_TAG, `receive error: id: ${msg.id}, code: ${msg.code}, error: ${msg.message}`);
+        switch (msg.code) {
+          case ErrorMessageCode.ChannelNotFound:
+            // 有 channelPath 说明该 channel 曾经被打开过
+            // 重新打开 channel
+            if (this.channelPath) {
+              // 暂停消息发送直到 server-ready
+              this.pause();
+              this.open(this.channelPath, this.clientId);
+            }
+            break;
+        }
+        break;
     }
   }
 
@@ -192,6 +161,9 @@ export class WSChannel {
    */
   open(path: string, clientId: string, connectionToken = randomString(16)) {
     this.channelPath = path;
+    this.clientId = clientId;
+
+    this.LOG_TAG = `[WSChannel id=${this.id} path=${path}]`;
 
     if (this.stateTracer.has(connectionToken)) {
       this.logger.warn(
@@ -202,15 +174,13 @@ export class WSChannel {
 
     this.stateTracer.record(connectionToken);
 
-    this.connection.send(
-      stringify({
-        kind: 'open',
-        id: this.id,
-        path,
-        clientId,
-        connectionToken,
-      }),
-    );
+    this.connection.send({
+      kind: 'open',
+      id: this.id,
+      path,
+      clientId,
+      connectionToken,
+    });
 
     if (this._ensureServerReady) {
       this.ensureOpenSend(path, clientId, connectionToken);
@@ -237,23 +207,19 @@ export class WSChannel {
   }
 
   send(content: string) {
-    this.inqueue(
-      stringify({
-        kind: 'data',
-        id: this.id,
-        content,
-      }),
-    );
+    this.inqueue({
+      kind: 'data',
+      id: this.id,
+      content,
+    });
   }
 
   sendBinary(data: Uint8Array) {
-    this.inqueue(
-      stringify({
-        kind: 'binary',
-        id: this.id,
-        binary: data,
-      }),
-    );
+    this.inqueue({
+      kind: 'binary',
+      id: this.id,
+      binary: data,
+    });
   }
   onError() {}
   close(code?: number, reason?: string) {
@@ -286,12 +252,25 @@ export class WSChannel {
     return conn;
   }
 
+  listen() {
+    return this.connection.onMessage((data) => {
+      this.dispatch(data);
+    });
+  }
+
   dispose() {
     if (this.timer) {
       clearTimeout(this.timer);
     }
     this.sendQueue = [];
     this._disposables.dispose();
+  }
+
+  ping() {
+    this.connection.send({
+      kind: 'ping',
+      id: this.id,
+    });
   }
 }
 
@@ -306,18 +285,16 @@ export class WSServerChannel extends WSChannel {
   messageQueue: ChannelMessage[] = [];
 
   clientId: string;
-  constructor(public connection: IConnectionShape<Uint8Array>, options: IWSServerChannelCreateOptions) {
+  constructor(public connection: IConnectionShape<ChannelMessage>, options: IWSServerChannelCreateOptions) {
     super(connection, options);
     this.clientId = options.clientId;
   }
   serverReady(token: string) {
-    this.connection.send(
-      stringify({
-        kind: 'server-ready',
-        id: this.id,
-        token,
-      }),
-    );
+    this.connection.send({
+      kind: 'server-ready',
+      id: this.id,
+      token,
+    });
   }
 
   dispatch(msg: ChannelMessage) {
@@ -328,128 +305,12 @@ export class WSServerChannel extends WSChannel {
       case 'binary':
         this.emitter.emit('binary', msg.binary);
         break;
+      case 'ping':
+        this.connection.send({
+          kind: 'pong',
+          id: this.id,
+        });
+        break;
     }
   }
 }
-
-export type ChannelMessage =
-  | PingMessage
-  | PongMessage
-  | OpenMessage
-  | ServerReadyMessage
-  | DataMessage
-  | BinaryMessage
-  | CloseMessage
-  | ErrorMessage;
-
-export const PingProtocol = Type.object('ping', {
-  clientId: Type.string(),
-  id: Type.string(),
-});
-
-export const PongProtocol = Type.object('pong', {
-  clientId: Type.string(),
-  id: Type.string(),
-});
-
-/**
- * `open` message is used to open a new channel.
- * `path` is used to identify which handler should be used to handle the channel.
- * `clientId` is used to identify the client.
- */
-export interface OpenMessage {
-  kind: 'open';
-  id: string;
-  path: string;
-  clientId: string;
-  connectionToken: string;
-}
-
-export const OpenProtocol = Type.object('open', {
-  clientId: Type.string(),
-  id: Type.string(),
-  path: Type.string(),
-  connectionToken: Type.string(),
-});
-
-/**
- * when server receive a `open` message, it should reply a `server-ready` message.
- * this is indicate that the channel is ready to use.
- */
-export interface ServerReadyMessage {
-  kind: 'server-ready';
-  id: string;
-  token: string;
-}
-
-export const ServerReadyProtocol = Type.object('server-ready', {
-  id: Type.string(),
-  token: Type.string(),
-});
-
-export enum ErrorMessageCode {
-  ChannelNotFound = 1,
-}
-
-export interface ErrorMessage {
-  kind: 'error';
-  id: string;
-  code: ErrorMessageCode;
-  message: string;
-}
-
-export const ErrorProtocol = Type.object('error', {
-  id: Type.string(),
-  code: Type.uint16(),
-  message: Type.string(),
-});
-
-export const DataProtocol = Type.object('data', {
-  id: Type.string(),
-  content: Type.string(),
-});
-
-export const BinaryProtocol = Type.object('binary', {
-  id: Type.string(),
-  binary: Type.binary(),
-});
-
-export const CloseProtocol = Type.object('close', {
-  id: Type.string(),
-  code: Type.uint32(),
-  reason: Type.string(),
-});
-
-const serializer = oneOf([
-  PingProtocol,
-  PongProtocol,
-  OpenProtocol,
-  ServerReadyProtocol,
-  DataProtocol,
-  BinaryProtocol,
-  CloseProtocol,
-  ErrorProtocol,
-]);
-
-export function stringify(obj: ChannelMessage): Uint8Array {
-  return serializer.serialize(obj);
-}
-
-export function parse(input: Uint8Array): ChannelMessage {
-  return serializer.deserialize(input) as any;
-}
-
-const _pingMessage: PingMessage = {
-  kind: 'ping',
-  id: '',
-  clientId: '',
-};
-
-const _pongMessage: PongMessage = {
-  kind: 'pong',
-  id: '',
-  clientId: '',
-};
-
-export const pingMessage = stringify(_pingMessage);
-export const pongMessage = stringify(_pongMessage);
